requirements.txt
pandas
numpy
pyarrow
pyyaml
obspy
cdflib
zarr
numcodecs
fastapi
uvicorn
jinja2
plotly
matplotlib
httpx
pytest
pytest-cov
ruff
black
